scripts/requirements.txt
pycryptodome>=3.20.0
textual>=6.7.0
pyperclip>=1.8.2
